requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
paho-mqtt==1.6.1
numpy==1.24.3
scipy==1.10.1
pydub==0.25.1
SpeechRecognition==3.10.0
googletrans==4.0.0-rc1
